pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lm-eval-so"
version = "0.2.0"
description = "LM Evaluation and Synthetic Data Orchestrator"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "LM-Eval-SO" }]
keywords = ["lm", "evaluation", "dataset", "runner", "evaluator", "synthetic"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Testing",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "PyYAML>=6.0",
  "openai>=1.35.0",
  "pydantic>=2.0",
  "fastmcp>=0.1.0",
]

# 이 루트 프로젝트는 주로 개발/문서/예제용 메타 패키지 역할을 하며,
# 실제 파이썬 패키지는 각 서브 디렉터리(src/lm_eval_so/generator, runner, evaluator)에 정의된다.

[tool.setuptools]
package-dir = {"" = "src"}

[project.scripts]
lm-eval-mcp = "lm_eval_so.mcp_server:main"
lm-eval-runner = "lm_eval_so.runner.cli:main"
lm-eval-evaluator = "lm_eval_so.evaluator.cli:main"
lm-eval-generator = "lm_eval_so.generator.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
